# perf: reduce the overhead incurred when deserializing Tuple