--- a/src/catalog/table.rs
+++ b/src/catalog/table.rs
@@ -1,11 +1,12 @@
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 use std::sync::Arc;
+use itertools::Itertools;
 
 use crate::catalog::{ColumnCatalog, ColumnRef};
 use crate::errors::DatabaseError;
 use crate::types::index::{IndexMeta, IndexMetaRef};
-use crate::types::ColumnId;
+use crate::types::{ColumnId, LogicalType};
 
 pub type TableName = Arc<String>;
 
@@ -56,6 +57,17 @@ impl TableCatalog {
 
     pub(crate) fn all_columns(&self) -> Vec<ColumnRef> {
         self.columns.values().map(Arc::clone).collect()
+    }
+
+    pub(crate) fn primary_key(&self) -> Result<(usize, &ColumnRef), DatabaseError> {
+        self.columns.iter().map(|(_, column)| column).enumerate().find(|(_, column)| column.desc.is_primary).ok_or(DatabaseError::PrimaryKeyNotFound)
+    }
+
+    pub(crate) fn types(&self) -> Vec<LogicalType> {
+        self.columns
+            .iter()
+            .map(|(_, column)| *column.datatype())
+            .collect_vec()
     }
 
     /// Add a column to the table catalog.