--- a/src/execution/volcano/dql/explain.rs
+++ b/src/execution/volcano/dql/explain.rs
@@ -1,12 +1,10 @@
 use crate::catalog::ColumnCatalog;
-use crate::catalog::ColumnRef;
 use crate::errors::DatabaseError;
 use crate::execution::volcano::{BoxedExecutor, ReadExecutor};
 use crate::planner::LogicalPlan;
 use crate::storage::Transaction;
 use crate::types::tuple::Tuple;
 use crate::types::value::DataValue;
-use crate::types::value::ValueRef;
 use futures_async_stream::try_stream;
 use std::sync::Arc;
 
@@ -29,8 +27,8 @@ impl<T: Transaction> ReadExecutor<T> for Explain {
 impl Explain {
     #[try_stream(boxed, ok = Tuple, error = DatabaseError)]
     pub async fn _execute(self) {
-        let columns: Vec<ColumnRef> = vec![Arc::new(ColumnCatalog::new_dummy("PLAN".to_string()))];
-        let values: Vec<ValueRef> = vec![Arc::new(DataValue::Utf8(Some(self.plan.explain(0))))];
+        let columns = Arc::new(vec![Arc::new(ColumnCatalog::new_dummy("PLAN".to_string()))]);
+        let values = vec![Arc::new(DataValue::Utf8(Some(self.plan.explain(0))))];
 
         yield Tuple {
             id: None,