--- a/src/execution/volcano/dql/seq_scan.rs
+++ b/src/execution/volcano/dql/seq_scan.rs
@@ -26,7 +26,7 @@ impl SeqScan {
     pub async fn _execute<T: Transaction>(self, transaction: &T) {
         let ScanOperator {
             table_name,
-            projection_columns: columns,
+            columns,
             limit,
             ..
         } = self.op;