### What problem does this PR solve?

- Only deserialize the values ​​that require project
- Use SchemaRef to avoid repeating collect_vec for ColumnRef

2_00_000 row: 190ms -> 142ms
```shell
FnckSQL: select * from t1                                                                                  
                        time:   [141.42 ms 142.85 ms 144.54 ms]
                        change: [-19.602% -17.850% -16.129%] (p = 0.00 < 0.05)
                        Performance has improved.

SQLite: select * from t1                                                                           
                        time:   [41.281 ms 42.103 ms 42.832 ms]
                        change: [-2.0020% -0.8158% +0.8000%] (p = 0.27 > 0.05)
                        No change in performance detected.

```

### Code changes

- [x] Has Rust code change
- [x] Has CI related scripts change

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [x] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

### Note for reviewer